move to workspace_aliases to internal header
